Bugfix: PBS log jobname
The function to produce PBS output logs references an undefined
variable, which is fixed in this bug.

Unfortuantely this info is being pulled from expt's config variable,
which is now effectively a common input argument.

There ought to be a better way to structure this, either adding a config
to the scheduler, or more likely the experiment should be pulling the
relevant info from the scheduler to produce its own log.

I am not yet sure the "proper" way to sort this, so I added an arguemnt
which is now unused in Slurm.  Will hopefully revisit this someday (but
probably not).